data type fixed

+++ Sensor/Sensor/Models/AirQualityData.cs
@@ -17,15 +17,15 @@ namespace Sensor.Models
         public DateTime Date { get; set; }
         public TimeSpan Time { get; set; }
         public float CO_GT { get; set; }
-        public int PT08_S1_CO { get; set; }
-        public int NMHC_GT { get; set; }
+        public float PT08_S1_CO { get; set; }
+        public float NMHC_GT { get; set; }
         public float C6H6_GT { get; set; }
-        public int PT08_S2_NMHC { get; set; }
-        public int NOx_GT { get; set; }
-        public int PT08_S3_NOx { get; set; }
-        public int NO2_GT { get; set; }
-        public int PT08_S4_NO2 { get; set; }
-        public int PT08_S5_O3 { get; set; }
+        public float PT08_S2_NMHC { get; set; }
+        public float NOx_GT { get; set; }
+        public float PT08_S3_NOx { get; set; }
+        public float NO2_GT { get; set; }
+        public float PT08_S4_NO2 { get; set; }
+        public float PT08_S5_O3 { get; set; }
         public float T { get; set; }
         public float RH { get; set; }
         public float AH { get; set; }
